+++ src/cvtools/datasets/classification/generic/classification_dataset_pt.py
@@ -70,7 +70,7 @@ class ClassificationDatasetPT(ClassificationDataset, Dataset):
 
         image, label = super().__getitem__(id)
         image = ToImage()(image)
-        label = torch.tensor(label, dtype=torch.float32)
+        label = torch.tensor(label)
 
         if self.transform:
             image = self.transform(image)
