+++ src/cvtools/datasets/classification/sun/sun_pt.py
@@ -68,7 +68,7 @@ class SUNDatasetPT(SUNDataset, Dataset):
         image, label = super().__getitem__(index)
 
         image = ToImage()(image)
-        label = torch.tensor(label, dtype=torch.float32)
+        label = torch.tensor(label)
 
         if self.transform:
             image = self.transform(image)
